% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dna.R
\name{add_dna}
\alias{add_dna}
\title{Define a dna reference genome to be used as additional database, or change an existing one.
If parameters are missing, a menu will guide the user choice. Otherwise, user input is used!}
\usage{
add_dna(ARG_LIST, dna.file = "", vcf.file = "", update = F)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments()}

\item{dna.file}{File in fasta format with DNA information}

\item{vcf.file}{File in vcf format with SNP information}

\item{update}{Renew existing data}
}
\description{
Define a dna reference genome to be used as additional database, or change an existing one.
If parameters are missing, a menu will guide the user choice. Otherwise, user input is used!
}
