% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_db.R
\name{add_local_db}
\alias{add_local_db}
\title{Add a RNA database from a local source into this project}
\usage{
add_local_db(ARG_LIST, db.name = NA, rna.file = NA, update = F)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments()}

\item{db.name}{Database name}

\item{rna.file}{File in gff3 or fasta format with RNA information}

\item{update}{Renew existing data}
}
\description{
Add a RNA database from a local source into this project
}
