% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_remote_db.R
\name{add_remote_db}
\alias{add_remote_db}
\title{Add a RNA database from a remote source into this project}
\usage{
add_remote_db(ARG_LIST, db.type = "ensembl", db.version = "", update = F)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments()}

\item{db.type}{Database type, either "refseq" or "ensembl" (default)}

\item{db.version}{Release number}

\item{update}{Renew existing data}
}
\description{
Add a RNA database from a remote source into this project
}
