% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a project directory}
\usage{
create_project(ARG_LIST)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments.R}
}
\description{
Create a project directory
}
