% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_db_usage.R
\name{define_db_usage}
\alias{define_db_usage}
\title{Define the database processing order}
\usage{
define_db_usage(ARG_LIST, ranking = NA)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments.R}

\item{ranking}{numeric vector with database processing (=rank) order}
}
\value{
updated ARG_LIST
}
\description{
Database processing order is defined by a numeric vector with one value for each available database DB1 to DBx.

Possible values:
\describe{
 \item{1}{most important}
 \item{2}{2nd most important}
 \item{x}{xth most important}
 \item{0}{excluded}
}
i.e. c(1,2,1,0) means DB1 and DB3 are processed first, then DB2, DB4 isn't used at all
}
