% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_db.R
\name{index_db}
\alias{index_db}
\title{Add bowtie index to an existing database}
\usage{
index_db(ARG_LIST, db.name)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments.R}

\item{db.name}{Database name}
}
\description{
Add bowtie index to an existing database
}
