% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_dbs.R
\name{list_available_dbs}
\alias{list_available_dbs}
\title{List all "installed" databases}
\usage{
list_available_dbs(ARG_LIST)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments.R}
}
\description{
List all "installed" databases
}
