% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_db.R
\name{map_db}
\alias{map_db}
\title{Map sequences in 'probes.fasta' against a custom database reference}
\usage{
map_db(ARG_LIST, db.name)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments.R}

\item{db.name}{Database name}
}
\description{
Map sequences in 'probes.fasta' against a custom database reference
}
