% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chip.R
\name{prepare_chip}
\alias{prepare_chip}
\title{Prepare chip files and create file "probes.fasta"}
\usage{
prepare_chip(ARG_LIST)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments.R}
}
\description{
Prepare chip files and create file "probes.fasta"
}
