% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare all input data for the revised probeset assignment}
\usage{
prepare_data(
  WORK_PATH,
  PROJ_NAME,
  CHIP_PATH,
  SPECIES,
  NCPU,
  STRICT_MAPPING = T,
  ASK_QUESTIONS = T
)
}
\arguments{
\item{WORK_PATH}{Working directory}

\item{PROJ_NAME}{Project name}

\item{CHIP_PATH}{Chip path with one pgf and one mps file}

\item{SPECIES}{Species name in latin language, i.e. "Sus_scrofa"}

\item{NCPU}{Number of threads to use}

\item{STRICT_MAPPING}{Avoid SNP correction of probes}

\item{ASK_QUESTIONS}{Answer questions automatically with their defaults}
}
\description{
The function wraps functions that can also be directly used in given order:
prepare_input_arguments, create_project, prepare_chip, add_local_db, add_remote_db, add_dna, index_db, map_db, define_db_usage
}
