% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_arguments.R
\name{prepare_input_arguments}
\alias{prepare_input_arguments}
\title{Prepare all input arguments for the revised probeset assignment}
\usage{
prepare_input_arguments(
  WORK_PATH,
  PROJ_NAME,
  CHIP_PATH,
  SPECIES,
  NCPU,
  STRICT_MAPPING,
  ASK_QUESTIONS
)
}
\arguments{
\item{WORK_PATH}{Working directory}

\item{PROJ_NAME}{Project name}

\item{CHIP_PATH}{Chip path with one pgf and one mps file}

\item{SPECIES}{Species name in latin language, i.e. "Sus_scrofa"}

\item{NCPU}{Number of threads to use}

\item{STRICT_MAPPING}{Avoid SNP correction of probes}

\item{ASK_QUESTIONS}{Answer questions automatically with their defaults}
}
\description{
Prepare all input arguments for the revised probeset assignment
}
