% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the revised probeset assignment}
\usage{
run(
  ARG_LIST,
  keep_unqualified_probesets = F,
  sep1 = "||",
  sep2 = ";",
  sep3 = ","
)
}
\arguments{
\item{ARG_LIST}{List with user settings created by prepare_input_arguments() or prepare()}

\item{keep_unqualified_probesets}{Keep unqualified probesets for better background correction and later analysis}

\item{sep1}{Separator for different database results at identical genomic position}

\item{sep2}{Separator for different results}

\item{sep3}{Separator for different gene.alternative results}
}
\description{
Run the revised probeset assignment
}
