% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_report.R
\name{show_report}
\alias{show_report}
\title{Show facts about the finished revision}
\usage{
show_report(ARG_LIST)
}
\arguments{
\item{ARG_LIST}{list with user settings created by prepare_input_arguments() or prepare_data()}
}
\description{
Show facts about the finished revision
}
